+++ packages/components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.6-alpha.1](https://github.com/gpn-prototypes/vega-ui/compare/@gpn-prototypes/vega-ui@2.15.6-alpha.0...@gpn-prototypes/vega-ui@2.15.6-alpha.1) (2020-12-01)
+
+**Note:** Version bump only for package @gpn-prototypes/vega-ui
+
+
+
+
+
+## [2.15.6-alpha.0](https://github.com/gpn-prototypes/vega-ui/compare/@gpn-prototypes/vega-ui@2.15.5...@gpn-prototypes/vega-ui@2.15.6-alpha.0) (2020-12-01)
+
+**Note:** Version bump only for package @gpn-prototypes/vega-ui
+
+
+
+
+
 ## [2.15.5](https://github.com/gpn-prototypes/vega-ui/compare/@gpn-prototypes/vega-ui@2.15.4...@gpn-prototypes/vega-ui@2.15.5) (2020-11-30)
 
 **Note:** Version bump only for package @gpn-prototypes/vega-ui